+++ blog/.env.template
@@ -1,0 +1,7 @@
+PUBLIC_APPWRITE_ENDPOINT=YOUR_ENDPOINT
+PUBLIC_APPWRITE_PROJECT_ID=YOUR_PROJECT_ID
+
+PUBLIC_APPWRITE_DB_ID=YOUR_DATABASE_ID
+PUBLIC_APPWRITE_POSTS_ID=YOUR_POSTS_COLLECTION_ID
+PUBLIC_APPWRITE_COMMENTS_ID=YOUR_COMMENTS_COLLECTION_ID
+PUBLIC_APPWRITE_BUCKET_ID=YOUR_BUCKET_ID
