+++ blog/.gitignore
@@ -1,5 +1,6 @@
 # build output
 dist/
+.vercel
 # generated types
 .astro/
