+++ blog/.vscode/extensions.json
