+++ blog/public/placeholder-hero.jpg
