+++ blog/src/components/Footer.astro
