+++ blog/src/components/SideNav.astro
@@ -25,7 +25,7 @@ import Account from "./Account.astro";
       <ul class="drop-list">
         <li class="drop-list-item">
           <NavLink
-            href="https://docs.astro.build/en/guides/baas/appwriteio/"
+            href="https://docs.astro.build/en/guides/backend/appwriteio/"
             target="_blank"
             rel="noopener noreferrer"
           >
