+++ blog/src/consts.ts
