+++ blog/src/env.d.ts
@@ -5,5 +5,7 @@ interface ImportMetaEnv {
   readonly PUBLIC_APPWRITE_ENDPOINT: string;
   readonly PUBLIC_APPWRITE_PROJECT_ID: string;
   readonly PUBLIC_APPWRITE_DB_ID: string;
-  readonly PUBLIC_APPWRITE_BLOG_ID: string;
+  readonly PUBLIC_APPWRITE_POSTS_ID: string;
+  readonly PUBLIC_APPWRITE_COMMENTS_ID: string;
+  readonly PUBLIC_APPWRITE_BUCKET_ID: string;
 }
