+++ blog/src/pages/blog/[...slug].astro
@@ -15,7 +15,7 @@ if (!slug) {
 
 const entries = (await appwriteDatabases.listDocuments(
   import.meta.env.PUBLIC_APPWRITE_DB_ID,
-  import.meta.env.PUBLIC_APPWRITE_BLOG_ID,
+  import.meta.env.PUBLIC_APPWRITE_POSTS_ID,
   [Query.equal("slug", slug)]
 )) as BlogPostList;
