+++ blog/src/pages/index.astro
@@ -11,7 +11,7 @@ import SiteLayout from "../layouts/SiteLayout.astro";
       rel="noopener noreferrer"
       style="text-decoration: underline">repo</a
     > for the guide found on <a
-      href="https://docs.astro.build/en/guides/baas/appwriteio/"
+      href="https://docs.astro.build/en/guides/backend/appwriteio/"
       target="_blank"
       rel="noopener noreferrer"
       style="text-decoration: underline">Astro's Doc Site</a
