--- example-blog/.env.template
@@ -1,6 +1,0 @@
-PUBLIC_APPWRITE_ENDPOINT=YOUR_ENDPOINT
-PUBLIC_APPWRITE_PROJECT_ID=YOUR_PROJECT_ID
-
-PUBLIC_APPWRITE_DB_ID=YOUR_DATABASE_ID
-PUBLIC_APPWRITE_BLOG_ID=YOUR_BLOG_COLLECTION_ID
-PUBLIC_APPWRITE_COMMENTS_ID=YOUR_BLOG_COMMENTS_COLLECTION_ID
