--- example-blog/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "example-blog",
-	"type": "module",
-	"version": "0.0.1",
-	"scripts": {
-		"dev": "astro dev",
-		"start": "astro dev",
-		"build": "astro build",
-		"preview": "astro preview",
-		"astro": "astro"
-	},
-	"dependencies": {
-		"@appwrite.io/pink": "^0.0.5",
-		"@astrojs/mdx": "^0.16.0",
-		"@astrojs/rss": "^2.1.0",
-		"@astrojs/sitemap": "^1.0.1",
-		"@astrojs/vercel": "^3.1.2",
-		"appwrite": "^10.2.0",
-		"astro": "^2.0.14",
-		"nanostores": "^0.7.3"
-	}
-}
