--- example-blog/src/lib/utilities.ts
@@ -1,1 +1,0 @@
-export const dateForDateTimeInputValue = (date:Date) => new Date(date.getTime() + new Date().getTimezoneOffset() * -60 * 1000).toISOString().slice(0, 19)
